Behaviour of repsTo last_success and last_attempt tests
This test validates last_success and last_attempt in repsTo. These attributes should contain
timestamps, but currently they are reset on every replication sync. This leads to
missing times visible in samba-tool drs showrepl.

As this is the test only, it is set to known fail, pending the following patches.

Signed-off-by: Dirk Godau <voidswitch@gmail.com>